requirements.txt
PyMySQL~=1.0.2
python-dotenv~=0.19.1
python-telegram-bot~=13.7
python-telegram-bot-calendar~=1.0.5
cryptography~=35.0.0